requirements.txt
langchain==1.2.6
langchain_community==0.4.1
langchain_core==1.2.7
langgraph==1.0.6
pydantic==2.12.5
pymilvus==2.3.4
python-dotenv==1.2.1
Requests==2.32.5
sentence_transformers==2.2.2
torch==2.1.2+cu118

